Cleaned up some stuff
