+++ src/hooks/index.ts
@@ -1,6 +1,5 @@
 export { useFilterInputChangeHandler } from "hooks/useFilterInputChangeHandler";
 export { useGetUserPatchesPageTitleAndLink } from "hooks/useGetUserPatchesPageTitleAndLink";
-export { useNotificationModal } from "hooks/useNotificationModal";
 export { useOnClickOutside } from "hooks/useOnClickOutside";
 export { useVersionTaskStatusSelect } from "hooks/useVersionTaskStatusSelect";
 export { useLegacyUIURL } from "hooks/useLegacyUIURL";
